requirements.txt
anthropic
boto3
chromadb # vector db
faiss-cpu>=1.12.0
flask
langchain
langchain_community
langchain-chroma
numpy>=2.2.6
openai
pandas>=2.3.2
PyMuPDF
# pypdf
pytest
python-dotenv>=1.1.1
requests
sentence_transformers
